pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "consite"
version = "0.1.0"
description = "ConSite: conserved-domain alignment and conserved-site visualization from protein FASTA"
authors = [{name = "Joey Wagner"}, {name = "Yang Li"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "biopython>=1.81",
  "pandas>=2.0",
  "numpy>=1.24",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "seaborn>=0.12",
]

[project.scripts]
consite = "consite.cli:main"